- Added ContactUs model for handling contact form submissions. - Created AppointmentWidget component for patient appointment navigation. -enhance UI/UX and fixed bugs in the front end.
