+++ client/src/app/models/contactUs.ts
@@ -1,0 +1,11 @@
+export class ContactUs {
+    name: string;
+    email: string;
+    message: string;
+
+    constructor(name: string, email: string, message: string) {
+        this.name = name;
+        this.email = email;
+        this.message = message;
+    }
+}
