+++ client/src/app/pages/patient/components/topbarwidget.component.ts
@@ -1,19 +1,22 @@
-import { Component } from '@angular/core';
+
+import { Component, OnInit } from '@angular/core';
 import { StyleClassModule } from 'primeng/styleclass';
-import { Router, RouterModule } from '@angular/router';
+import { Router, RouterModule, NavigationEnd } from '@angular/router';
 import { RippleModule } from 'primeng/ripple';
 import { ButtonModule } from 'primeng/button';
-import { AppFloatingConfigurator } from "../../admin/components/app.floatingconfigurator";
+import { AppFloatingConfigurator } from '../../admin/components/app.floatingconfigurator';
 import { AuthStateService } from '../../../service/auth-state.service';
 import { ToastModule } from 'primeng/toast';
 import { MessageService } from 'primeng/api';
 import { MenubarModule } from 'primeng/menubar';
 import { NotificationService } from '../../../service/notification.service';
+import { filter } from 'rxjs/operators';
+import { CommonModule } from '@angular/common';
 
 @Component({
     selector: 'patient-topbar-widget',
     providers: [MessageService],
-    imports: [MenubarModule, RouterModule, StyleClassModule, ButtonModule, RippleModule, AppFloatingConfigurator, ToastModule],
+    imports: [CommonModule, MenubarModule, RouterModule, StyleClassModule, ButtonModule, RippleModule, AppFloatingConfigurator, ToastModule],
     template: `<a class="flex items-center" href="/patient">
             <svg viewBox="0 0 54 40" fill="none" xmlns="http://www.w3.org/2000/svg" class="h-12 mr-2">
                 <path
@@ -42,47 +45,99 @@ import { NotificationService } from '../../../service/notification.service';
         <div class="items-center dark:bg-surface-900 grow justify-between hidden lg:flex absolute lg:static w-full left-0 top-full px-12 lg:px-0 z-20 rounded-border dark:bg-black">
             <ul class="list-none p-0 m-0 flex lg:items-center select-none flex-col lg:flex-row cursor-pointer gap-8">
                 <li>
-                    <a (click)="router.navigate(['/patient'], { fragment: 'hero' })"  pRipple class="px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl">
+                    <a (click)="navigateTo('/patient', 'hero')" pRipple [ngClass]="{ 'px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl': true, 'active-tab': isActiveRoute('/patient') && currentFragment === 'hero' }">
                         <span>Home</span>
                     </a>
                 </li>
                 <li>
-                    <a (click)="router.navigate(['/patient'], { fragment: 'highlights' })" pRipple class="px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl">
+                    <a (click)="navigateTo('/patient', 'highlights')" pRipple [ngClass]="{ 'px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl': true, 'active-tab': isActiveRoute('/patient') && currentFragment === 'highlights' }">
                         <span>Features</span>
                     </a>
                 </li>
                 <li>
-                    <a (click)="router.navigate(['/patient'], { fragment: 'aboutus' })" pRipple class="px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl">
+                    <a (click)="navigateTo('/patient', 'aboutus')" pRipple [ngClass]="{ 'px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl': true, 'active-tab': isActiveRoute('/patient') && currentFragment === 'aboutus' }">
                         <span>AboutUs</span>
                     </a>
                 </li>
                 <li>
-                    <a (click)="router.navigate(['/patient'], { fragment: 'contactus' })" pRipple class="px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl">
+                    <a (click)="navigateTo('/patient', 'contactus')" pRipple [ngClass]="{ 'px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl': true, 'active-tab': isActiveRoute('/patient') && currentFragment === 'contactus' }">
                         <span>ContactUs</span>
                     </a>
                 </li>
                 <li>
-                    <a (click)="router.navigate(['/patient/appointment'])" pRipple class="px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl">
-                        <span>Appointment</span>
-                    </a>
+                    <div class="relative group inline-block">
+                        <a pRipple [ngClass]="{ 'px-0 py-4 text-surface-900 dark:text-surface-0 font-medium text-xl cursor-pointer': true, 'active-tab': isActiveRoute('/patient/appointment') }">
+                            <span>Appointment</span>
+                        </a>
+
+                        <!-- Dropdown -->
+                        <div class="absolute hidden group-hover:block bg-white dark:bg-gray-800 shadow-lg rounded-md mt-2 z-50 w-48">
+                            <a
+                                (click)="router.navigate(['/patient/appointment'])"
+                                [ngClass]="{
+                                    'block px-4 py-2 text-gray-800 dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700': true,
+                                    'active-dropdown-item': isActiveRoute('/patient/appointment') && !isActiveRoute('/patient/appointment/view')
+                                }"
+                            >
+                                Make Appointments
+                            </a>
+                            <a
+                                (click)="router.navigate(['/patient/appointment/view'])"
+                                [ngClass]="{ 'block px-4 py-2 text-gray-800 dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700': true, 'active-dropdown-item': isActiveRoute('/patient/appointment/view') }"
+                            >
+                                View Appointments
+                            </a>
+                        </div>
+                    </div>
                 </li>
-                <!-- <li>
-                    <p-menubar [model]="nestedMenuItems"> </p-menubar>
-                </li> -->
             </ul>
             <div class="flex border-t lg:border-t-0 border-surface py-4 lg:py-0 mt-4 lg:mt-0">
                 <button pButton pRipple label="Logout" [rounded]="true" (click)="logout()"></button>
                 <app-floating-configurator />
             </div>
-        </div> `
+        </div>
+
+        <style>
+            .active-tab {
+                color: var(--primary-color) !important;
+                border-bottom: 2px solid var(--primary-color);
+            }
+            .active-dropdown-item {
+                background-color: rgba(var(--primary-color-rgb), 0.1);
+                color: var(--primary-color) !important;
+            }
+        </style>`
 })
-export class TopbarWidget {
+export class TopbarWidget implements OnInit {
+    currentUrl: string = '';
+    currentFragment: string | null = null;
+
     constructor(
         public router: Router,
         private authStateService: AuthStateService,
         private messageService: MessageService,
-         private notificationService: NotificationService
+        private notificationService: NotificationService
     ) {}
+
+    ngOnInit() {
+        // Initialize the current URL
+        this.currentUrl = this.router.url.split('#')[0];
+        this.currentFragment = this.router.url.includes('#') ? this.router.url.split('#')[1] : null;
+
+        // Subscribe to router events to update active tab
+        this.router.events.pipe(filter((event) => event instanceof NavigationEnd)).subscribe((event: any) => {
+            this.currentUrl = event.url.split('#')[0];
+            this.currentFragment = event.url.includes('#') ? event.url.split('#')[1] : null;
+        });
+    }
+
+    navigateTo(route: string, fragment: string) {
+        this.router.navigate([route], { fragment: fragment });
+    }
+
+    isActiveRoute(route: string): boolean {
+        return this.currentUrl.startsWith(route);
+    }
 
     logout() {
         this.authStateService.clear();
@@ -91,26 +146,4 @@ export class TopbarWidget {
             this.router.navigate(['/auth/login']);
         }, 2000);
     }
-
-    // nestedMenuItems = [
-    //     {
-    //         label: 'Appointments',
-    //         items: [
-    //             {
-    //                 label: 'Add Appointments',
-    //                 icon: 'pi pi-calendar-plus',
-    //                 command: () => {
-    //                     this.router.navigate(['/patient/appointment/add']);
-    //                 }
-    //             },
-    //             {
-    //                 label: 'View Appointments',
-    //                 icon: 'pi pi-eye',
-    //                 command: () => {
-    //                     this.router.navigate(['/patient/appointment/view']);
-    //                 }
-    //             }
-    //         ]
-    //     }
-    // ];
 }
